Implement MNIST data loading and preprocessing functionality
- Added data_loader.go to handle downloading, decompressing, and loading the MNIST dataset.
- Updated mnist_example.go to utilize the new loading functions and streamline data handling.
- Modified .gitignore to ensure MNIST data files are properly ignored.
